# docs/api/openapi.v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "RKM Cinema API",
    "version": "2.0",
    "description": " (FROZEN v1 contract \u2014 see docs/adr/ADR-0001; additive-only)"
  },
  "paths": {
...
    "/api/auth/login": {
      "post": {
        "summary": "Login",
        "description": "Exchange Jellyfin credentials for an app session (and nothing else).\n\nIdentity is DELEGATED: the username/password go to Jellyfin's\n``/Users/AuthenticateByName`` and are never stored; what we keep is the access\ntoken Jellyfin handed back, behind an opaque id this app invented.",
        "operationId": "login_api_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/auth/profile": {
      "post": {
        "summary": "Select Profile",
        "description": "Switch this session to a profile \u2014 the whole of \"who is watching\".\n\nWhy it can only work this way: **Jellyfin has no impersonation** (measured 2026-09-12 \u2014 the\nonly user-scoped token endpoint is ``/Users/AuthenticateByName``), so the app asks Jellyfin to\nauthenticate AS the chosen profile. A password-less profile signs in with a blank password; a\nprotected one needs its own. **The administrator knowing a member's password is not the\nmechanism \u2014 resetting it from Settings \u2192 Household is**, which is why that rail exists.\n\nSelecting the ADMINISTRATOR'S OWN profile requires the administrator's password: the device\nholds their session, so this is what stops a guest on the shared iPad from walking into it\n(decision 3, 2026-09-12). Switching back is therefore always deliberate, never automatic.",
        "operationId": "select_profile_api_auth_profile_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SelectProfileRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SelectProfileResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LoginRequest": {
        "properties": {
          "username": {
            "type": "string",
            "title": "Username",
            "default": ""
          },
          "password": {
            "type": "string",
            "title": "Password",
            "default": ""
          }
        },
        "type": "object",
        "title": "LoginRequest",
        "description": "Sign in as a Jellyfin user (AUTH_MULTIUSER_PLAN \u00a73.7, Phase 0).\n\nIdentity is DELEGATED: these credentials are used once, against Jellyfin's\n``/Users/AuthenticateByName``, and never stored. What the app keeps is the access\ntoken Jellyfin returned, behind an opaque id it invented itself."
      },
      "LoginResponse": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok",
            "default": true
          },
          "user": {
            "$ref": "#/components/schemas/SessionUser"
          },
          "expires": {
            "type": "string",
            "title": "Expires",
            "default": ""
          }
        },
        "type": "object",
        "title": "LoginResponse"
      },
...
      "ProfileUser": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "default": ""
          },
          "name": {
            "type": "string",
            "title": "Name",
            "default": ""
          },
          "is_admin": {
            "type": "boolean",
            "title": "Is Admin",
            "default": false
          },
          "has_password": {
            "type": "boolean",
            "title": "Has Password",
            "default": false
          },
          "disabled": {
            "type": "boolean",
            "title": "Disabled",
            "default": false
          },
          "last_login": {
            "type": "string",
            "title": "Last Login",
            "default": ""
          }
        },
        "type": "object",
        "title": "ProfileUser",
        "description": "One selectable profile, as the \"Who's watching?\" picker needs it.\n\nDeliberately carries no credential and no token: a profile's password is only ever *asked*\nfor, never returned (PLEX_PROFILE_AUTH_PLAN \u00a74)."
      },
...
      "SelectProfileRequest": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id",
            "default": ""
          },
          "password": {
            "type": "string",
            "title": "Password",
            "default": ""
          }
        },
        "type": "object",
        "title": "SelectProfileRequest",
        "description": "``POST /api/auth/profile`` \u2014 switch this session to a profile.\n\n``password`` is that profile's own password (its Jellyfin credential): optional for a\npassword-less profile, **required for the administrator's own profile** so a shared device\ncannot walk into it (decision 3, 2026-09-12)."
      },
      "SelectProfileResponse": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok",
            "default": true
          },
          "profile": {
            "$ref": "#/components/schemas/ProfileUser"
          },
          "libraries": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Libraries"
          }
        },
        "type": "object",
        "title": "SelectProfileResponse",
        "description": "The profile now in effect, and the libraries it may see."
      },
      "SessionUser": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "default": ""
          },
          "name": {
            "type": "string",
            "title": "Name",
            "default": ""
          }
        },
        "type": "object",
        "title": "SessionUser",
        "description": "The signed-in user, as much of them as the browser is allowed to see."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}